Base Cosi ASDF system from @DavidMcLain's current work
Needs the subsequent patches containing the following ASDF system to
load:

    data-objects.diff
    useful-macros.diff
    mpcompat.diff
    crypto.diff
    lisp-object-encoder.diff
    sdle-store.diff

This patch is a "shallow copy" of the original source in the sense
that it doesn't have the revision history of the files included.  A
large part of that history, potentially spanning decades,
i.e. anything before

    user:        David McClain <dbm@refined-audiometrics.com>
    date:        Wed Jan 17 16:09:58 2018 -0700
    summary:     update to Treaps

was not available to be used for history.