It's working! (solicit (first (listify-nodes)) :count-alive) returns correct result [5] most of the time on a graph with 5 nodes.
Added hopcount. It's essential to make waiting time decrease as hopcount increases,
because otherwise everybody's wait loop stops at essentially the same time and
lots of otherwise-valid replies get ignored. That's because in this protocol,
a non-reply conveys information unto itself. We might want to modify this to
bounce back a reply immediately in some cases (e.g. for :count-alive, return
0 for ignored requests). This would make the protocol run faster without so
much waiting, but at the cost of more communication.

Added interactive-logging-function to make it possible to see results in
listener instead of having to look at the *log* every time.

Ensured that initial node called with solicit would record that it had seen the
message. We weren't doing this before, and I haven't seen any evidence that it's
essential, but it's important to be consistent.