Gossip rebased on dev b40ef75 (#47)
* Move files here

* More code. Not quite there yet.

* More code. Add package and .asd file.

* Now it compiles. Still unfinished.

* Interim checkin

* Interim checkin

* Starting to build the graph

* Cleaned up the svg output -- but now it might be too big.
Removed an inspect call.

* About ready to try count-alive.

* Simulation of :COUNT-ALIVE beginning to run. Badly.

* Almost working

* Several improvements. Can now save and restore graphs for debugging.
We're getting replies we didn't expect; still to be debugged.

* Improve logging so we can see more info about what's going on.

* Added graphviz to asd file.

* It's working! (solicit (first (listify-nodes)) :count-alive)
returns correct result [5] most of the time on a graph with 5 nodes.

Added hopcount. It's essential to make waiting time decrease as hopcount increases,
because otherwise everybody's wait loop stops at essentially the same time and
lots of otherwise-valid replies get ignored. That's because in this protocol,
a non-reply conveys information unto itself. We might want to modify this to
bounce back a reply immediately in some cases (e.g. for :count-alive, return
0 for ignored requests). This would make the protocol run faster without so
much waiting, but at the cost of more communication.

Added interactive-logging-function to make it possible to see results in
listener instead of having to look at the *log* every time.

Ensured that initial node called with solicit would record that it had seen the
message. We weren't doing this before, and I haven't seen any evidence that it's
essential, but it's important to be consistent.

* Minor mods to ensure it will work on larger networks.
Now gives nearly-correct result on 100-node network for :COUNT-ALIVE.

* Wrote "accept-msg?" Haven't fully broken it out yet but I'm about to.
Everything still works.

* Accept-msg? is now being called in deliver-msg as it should have been all along, so we
don't create a new thread just to ignore a message.
However, things are screwy now and debugger frames are all messed up as though
there's a memory error.

* Broke the contract of "accept-msg?" Duh. Fixed.

* Set up message-cache in deliver-msg to avoid a race condition. Must both check
AND set this flag in the same thread!

* Git rid of some no-longer-relevant debugging code.

* Count-alive now returns a list of the live nodes, rather than just their number.
This way we can deduce which ones didn't respond.

* Little bit of tuning

* Start getting rid of ccl dependencies. Use mpcompat and uiop.
Start to make it work in Lispworks.

* Get rid of remaining CCL dependencies.

* Remove some reader-macro chars that confused LW.

* Add announce method.
Add readme.txt.

* Changed (most) gethash calls to kvs calls.

* Lookup-key message is done and tested.

* Minor simplification in count-alive

* Make it work in command-line CCL too.