Fix ASDF dependencies
We include only the package definitions in the EMOTIQ system without
any direct dependencies.  Otherwise, one will quickly end up with
recursive definitions that don't compile or, worse, systems whose
behavior is dependent on unreproducible ASDF load ops if all systems
merely :DEPEND-ON.  This problem was evidenced in the delivery code
Paul was working on, so this is the minimal pre-emptive modification
to get into `dev` to disturb developers' productive lives as little as
possible.