add witness slots to block and various refinements and corrections
Replace validator-keys-joining and validator-keys-leaving slots with three new slots: public-keys-of-witnesses, witness-bitmap, and block-signature. Moving towards preference for witness over validator as the term of choice. A couple of methods, ith-witness-signed-p and set-ith-witness-signed-p, abstract the bitmap implementation details. 

Fixed deficiency that TRANSACTIONS slot, although specified as a sequence, was assumed to be a list in compute-merkle-root-hash; now that is fixed. Actually, it would be preferrable for this to be a vector in most cases, but let's leave it open for now.

Hash-block needed to be updated to account for the witness slots. I believe hashing only the signature should be sufficient. Also decided in consultation with David that the serialization could be simplified by using lisp object encoder (loenc) on slots and no use or need for counting number of octets in the serialized input (c.f., Bitcoin's '80 byte header').

Removed comment about switching to Hash160 for TXID-like hash, since it was mistaken: that's not the hash used for TXID's in Bitcoin.