Use a test harness external to Travis CI file
Allows running the test-harness by the developer before committing to
Travis.

Behavior is as before: each test is run in a separate Lisp process,
tests stop on the first unsucessful test.

Each test invocation is now wrapped in a handler so erroring tests are
caught before Travis CI times out.

Add `ccl` to test coverage, as we are using it in the initial testnet
deployment.