Add initial networking benchmark suite socket-actors-bench.lisp
This file defines a simple transport-level benchmark called
'sendrecv-tcp-1' that opens a set of client-server socket pairs and
sends messages from the client to the server.

The basic benchmark function GOSSIP::BENCHMARK-SENDRECV-TCP-1 takes
keyword parameters to control how the test is executed. This is
especially meant to test the performance of different socket polling
strategies, message serialization methods, etc, with different levels
of load.

The higher-level function GOSSIP::SUITE-SENDRECV-TCP-1 can execute a
complete test campaign for the permutations of given variable values.
The results are formatted as CSV for separate analysis.

Notes:

- This benchmark is operating at the socket layer and does not
  exercise higher-level functionality from socket-actors and gossip.
- The benchmark runs all clients and servers on a single Lisp thread.
- This initial implementation has not yet been sanity-checked with
  tools like system call profilers and so may contain egregious bugs.