socket-actors.lisp: Switch to unified "select loop" [not debugged]
Switch socket-actors to using a single process to poll for data from
all peers. (This replaces the previous model of polling each
connection with a separate thread.)

The process waits on all open connections using:

  (USOCKET:WAIT-FOR-INPUT :READY-ONLY T :TIMEOUT 0.01)

The timeout means that the select is restarted at regular intervals to
pick up newly opened connections.

This code is not yet debugged!