Sketch changes to gossip.lisp to use GOSSIP/TRANSPORT interface
This is a preliminary (won't compile...) attempt to make gossip.lisp
use the GOSSIP/TRANSPORT interface. The intention is to see the
approximate impact of this change and whether or not we like the
transport interface as it is currently defined.

The main impact on gossip.lisp is to remove the code that is directly
related to sockets and networking. The rest - automatic creation of
proxy nodes, etc - is the same.

Separate further work is both to make these changes compile and run,
and also to take the backend of the networking code from
socket-actors.lisp and move that into gossip-transport.lisp.