Define defstub - used to still compiler re many unimplemented functions.
This commit is primarily intended to eliminate pesky warnings.

For each function X in the crypto library that was not defined that we
got complaints about being called but not defined, there's now a
(defstub X) form.  This makes sure it's defined, although calling it
would not do anything really useful.  It also makes it known that the
function is a stub by adding a a property to the function name. That
property gets used by the little WITH-FAST-IMPL facility to avoid
generating a call to it.  I also cleaned up the error handling in
WITH-FAST-IMPL: previously, it would silently trap any error, and then
call makeunbound.  That's not a very clean way to handle it. First, it
should at least print the error, and help keep track of it.  Second,
it should not use makunbound. Instead, it now puts a flag property on
the function name and uses that as the gate to decide whether to call
the "fast" function.

NB: the speed of checking fboundp vs. property checks should be about
the same or noise level difference in most Lisps.  However, I suspect
either check might be adding significant overhad to the raw "fast"
call; review at a later date when we do optimization.