Fix dropping into CCL debugger in GOSSIP:PING-OTHER-HOSTS
Guard multi-threaded creation of keypairs via a simple mutex.  Without
this patch, the invocation of GOSSIP:PING-OTHER-HOSTS with enough
hosts would fail as more than one thread would attempt to create the
hmac keypair for the socket connection simultaneously.

We proactively fix similar code in COSI-BLS which would be suceptible
to the same sort of bug if it were potentially called by many threads.