EMOTIQ:MAIN now runs gossip, REST, and websocket modes
Any putative main loop must not stop executing if a condition is
signalled.  The ultimate strategy is to provide meaningful restarts
that may be selected at runtime, but our current need is to not bomb
out another subsystem just because your code is broken.

Wrap any errors created by startup.