Fix cloaked simulator by explicitly setting a prepare timeout
When running the cloaked version of the simulator, each node must run
the computational intensive Bulletproof range proofs in order to
complete the prepare phase.  If the timeout in seconds specified by
COSI-SIMGEN::*COSI-PREPARE-TIMEOUT* isn't "long enough", a new
election will be held which implements the outstanding prepare
phase. A value of 40 ("40 seconds") seems to be long enough for a
simulation network of four nodes.