Clean up initial commit of new transactions around mempool clearing.
Added remove-transaction-from-mempool, which knows to use
get-hash-key-of-transaction to get its key. That eliminates whole
reason for the bug earlier (while preparing the previous commit) of
using the old-transactions-style key (completely my fault), which
required the painful and clever coding workaround that PT figured out,
essentially rehashing both the transactions in the block and the
transactions in the mempool and then comparing their hashes using
hash=.

Hopefully, it's clearer now with this change.  Added an error message
by node-check-transaction in the case of *newtx-p* just for emphasis.

To test: run with newtx mode, see the right number of Tx's (4)
dumped. To test old: restart, and run with old tx mode, see right
number Tx's (2) eventually built.