154675977-smart-contract-transactions-21june2018: Auto stash before revert of "unstub check-block-transactions, move it to block, implement transaction wait function"
Added a function that waits for transactions with options for timing out, printing ticks to show progress, and so on.  It's not at all on the production, but on the debugging-tracing side of things. Adds a bit of deeper insight what's going on down below, and beats waiting the full 60 seconds for four transactions to finish with no signs of life.