changes for fees and collect transaction, with other cleanups, fixes
Working collect transactions, as well as cleanup of several other things.

Code to support making new collect type transactions, including code
to compute the fee on a transaction.

Made up a new type for the initial coinbase transaction in the genesis
block: :coinbase

Improved code to dump transactions, leaving out the useless
pseudo/dummy inputs in the case of collect and coinbase
transactions, and noting the transaction type in the case of
unusual, i.e., non-spend type, transaction.

Wait-for-tx-count, used by demo/test code in node-sim, had
various changes in response to the new types of transactions. Now
it counts only spend transactions by default. There's an option
to override/specify which types of transaction are counted.

The node-sim file's run-new-tx demo now advises that it's waiting
for 3 spend transactions, and then that it has finished 3 spend
transactions.  At the end, the dumper shows all the types of
transactions.  The dumper now identifies specially any
transaction that is not of type :spend, which is the most common,
standard, and default type.