Overhaul test infrastructure to return nil status on failure
Depends on ASDF-TEST-HARNESS from the Emotiq Quicklisp distribution.

Consistently name test packages and directories 'test'.

`etc/test-harness.bash` works with new system under CCL.  We need to
revisit the script invocation for developer, but the CI should work
for all the CL implementations we test as it has a script which
normalizes the command line options for the `--load` and `--eval`
options.