Following is support for #'hello API method, which required some re-architecting of proxies etc.
Add *subpath* variable to filesystem.lisp to make it easier to
test on multiple Lisp instances on one machine.

Remove remaining references to stakes in gossip. Check ports
as well as addresses when setting up *hosts* in
processes talking to each other on the same hardware, to make
debugging easier.)

Add #'remote-monads. Improve #'transport-message-received to
not make a proxy except for incoming solicitations that expect
replies (and do it right this time).

Set *actual-tcp-gossip-port* properly, and use it properly.