Create top-level TRANSPORT networking module
TRANSPORT is the new top-level name for the GOSSIP/TRANSPORT
networking code. This code is actually independent of GOSSIP - they
integrate via parameterizable callback functions - and so making it an
independent module is simpler.

TRANSPORT/BLOCKING is the backend based on threads and blocking
operations via the USOCKETS library.

TRANSPORT/ASYNC is the backend based on a non-blocking event loop with
cl-async (libuv).

TRANSPORT/TEST contains some testing/benchmarking utility functions.
These can be used interactively or combined with a test harness (not
included.)